fix: fix int

+++ ranking-api/cmd/db/gorm.go
@@ -12,7 +12,7 @@ type Database struct {
 }
 
 type Ranking struct {
-	ID      int64  `gorm:"primaryKey"`
+	ID      int    `gorm:"primaryKey"`
 	Name    string `gorm:"not null"`
 	Phone   string `gorm:"unique;not null"`
 	Current int    `gorm:"not null"`
